ditto
svn:sync-xref-src-repo-rev: r69341

+++ src/VBox/VMM/VMMR3/DBGFReg.cpp
@@ -1,4 +1,4 @@
-/* $Id: DBGFReg.cpp 35469 2011-01-10 17:20:07Z knut.osmundsen@oracle.com $ */
+/* $Id: DBGFReg.cpp 35470 2011-01-10 17:20:40Z knut.osmundsen@oracle.com $ */
 /** @file
  * DBGF - Debugger Facility, Register Methods.
  */
@@ -114,7 +114,7 @@ typedef struct DBGFREGSET
     /** Array of lookup records. */
     struct DBGFREGLOOKUP   *paLookupRecs;
     /** The number of lookup records. */
-    size_t                  cLookupRecs;
+    uint32_t                cLookupRecs;
 
     /** The register name prefix. */
     char                    szPrefix[1];
